fix: move functions from the bricks component, to avoid circular dependencies
