+++ components/polylith/bricks/__init__.py
@@ -1,9 +1,7 @@
-from polylith.bricks.base import create_base, get_bases_data
-from polylith.bricks.component import create_component, get_components_data
+from polylith.bricks.base import create_base
+from polylith.bricks.component import create_component
 
 __all__ = [
     "create_base",
     "create_component",
-    "get_bases_data",
-    "get_components_data",
 ]
