+++ components/polylith/bricks/base.py
@@ -1,7 +1,5 @@
 from pathlib import Path
-from typing import List
 
-from polylith.bricks import component
 from polylith.bricks.brick import create_brick
 from polylith.repo import bases_dir
 from polylith.test import create_test
@@ -13,7 +11,3 @@ def create_base(path: Path, options: dict) -> None:
 
     create_brick(path, base_options)
     create_test(path, base_options)
-
-
-def get_bases_data(path: Path, ns: str) -> List[dict]:
-    return component.get_components_data(path, ns, bases_dir)
