+++ components/polylith/dirs/__init__.py
@@ -1,3 +1,3 @@
-from polylith.dirs.dirs import create_dir
+from polylith.dirs.dirs import create_dir, get_bases_data, get_components_data
 
-__all__ = ["create_dir"]
+__all__ = ["create_dir", "get_bases_data", "get_components_data"]
